creating basic mcmlln tools class
