+++ core/cleanup/cleanup.rb
@@ -1,14 +1,14 @@
 require_relative '../header.rb'
 require_relative '../metadata.rb'
 
-data_hash = Bkmkr::Tools.readjson(Metadata.configfile)
+data_hash = Mcmlln::Tools.readjson(Metadata.configfile)
 project_dir = data_hash['project']
 stage_dir = data_hash['stage']
 
 # Delete all the working files and dirs
-Bkmkr::Tools.deleteDir(Bkmkr::Paths.project_tmp_dir)
-Bkmkr::Tools.deleteFile(Bkmkr::Project.input_file)
-Bkmkr::Tools.deleteFile(Bkmkr::Paths.alert)
+Mcmlln::Tools.deleteDir(Bkmkr::Paths.project_tmp_dir)
+Mcmlln::Tools.deleteFile(Bkmkr::Project.input_file)
+Mcmlln::Tools.deleteFile(Bkmkr::Paths.alert)
 
 # LOGGING
